login-register logic/still needs a universal logic for every page so it will not be passed explicity the session object
