--- .idea/sqldialects.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/prisma/migrations/20250629021846_add_user_model/migration.sql" dialect="GenericSQL" />
-    <file url="file://$PROJECT_DIR$/prisma/migrations/20250701214121_init_favorites/migration.sql" dialect="GenericSQL" />
-  </component>
-</project>
