+++ package.json
@@ -14,10 +14,13 @@
     "@fortawesome/free-regular-svg-icons": "^6.7.2",
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
+    "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^6.11.0",
     "@reduxjs/toolkit": "^2.8.2",
+    "bcryptjs": "^3.0.2",
     "lodash": "^4.17.21",
     "next": "15.3.3",
+    "next-auth": "^4.24.11",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-redux": "^9.2.0",
