+++ src/app/(protected)/favorites/components/favoriteMovieCardComponent.tsx
