+++ src/app/(protected)/movies/styles/searchBar.css
