+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,37 @@
+import NextAuth, { AuthOptions } from 'next-auth';
+import CredentialsProvider from 'next-auth/providers/credentials';
+import { prisma } from '@/lib/prisma';
+import { PrismaAdapter } from '@next-auth/prisma-adapter';
+import bcrypt from 'bcryptjs';
+
+export const authOptions: AuthOptions = {
+  adapter: PrismaAdapter(prisma),
+  providers: [
+    CredentialsProvider({
+      name: 'Credentials',
+      credentials: {
+        email: { label: 'Email', type: 'email' },
+        password: { label: 'Password', type: 'password' },
+      },
+      async authorize(credentials) {
+        if (!credentials?.email || !credentials?.password) {return null;}
+        const user = await prisma.user.findUnique({
+          where: { email: credentials.email },
+        });
+        if (!user) {return null;}
+        const isValid = await bcrypt.compare(credentials.password, user.password);
+        if (!isValid) {return null;}
+        console.log(user.id, user.email);
+        return { id: user.id, email: user.email };
+      },
+    }),
+  ],
+  session: { strategy: 'jwt' },
+  secret: process.env.NEXTAUTH_SECRET,
+  pages: {
+    signIn: '/auth/login',
+  },
+};
+
+const handler = NextAuth(authOptions);
+export { handler as GET, handler as POST };
