--- src/app/layout/footer/footer.tsx
@@ -1,9 +1,0 @@
-const Footer = () => {
-  return (
-    <footer className="bg-blue-600 text-white p-4 text-center w-full mt-auto">
-      <p>&copy; Created from Nomorelife with &hearts;. 2025 All rights reserved.</p>
-    </footer>
-  );
-};
-
-export default Footer;
